Split DocumentFolderService → DocumentService + CorpusObjsService
Issue #N/A — PR review of opencontractserver/mcp/tools.py surfaced a
manual two-step fusion (`corpus.get_documents()` + `Document.objects
.visible_to_user(user).get(...)`) that silently re-derived a different
permission semantic than the canonical
`DocumentFolderService.get_corpus_documents`. The pattern was repeated
across MCP, LLMs, GraphQL, REST views, and util modules.

This commit refactors the service layer along the natural seam:

- DocumentService (opencontractserver/documents/document_service.py) is
  the source of truth for document-level operations and permissioning
  (creation, quota, validation, standalone lookup, document permissions).
- CorpusObjsService (opencontractserver/corpuses/corpus_objs_service.py)
  is the source of truth for corpus-scoped operations ("give me X inside
  corpus Y for user Z"). It owns documents-in-corpus today; designed to
  grow to annotations, notes, relationships, extracts, analyses, etc.

DocumentFolderService is preserved as a multiple-inheritance subclass of
both new services so the ~18 existing importers continue to work
unchanged. Cleanup tracked under follow-up issue.

New convenience methods on CorpusObjsService eliminate the buggy fusion:

- get_corpus_document_by_slug(user, corpus, slug)
- get_corpus_document_by_id(user, corpus, document_id)
- is_document_in_corpus(user, corpus, document_id)

All three are corpus-as-gate and IDOR-safe — same Document.DoesNotExist
fires whether the document doesn't exist, isn't in this corpus, or the
user lacks corpus READ.

Migrated 15 user-context call sites:
- opencontractserver/mcp/tools.py (5 sites)
- opencontractserver/mcp/resources.py (2 sites)
- opencontractserver/llms/agents/core_agents.py (2 sites)
- opencontractserver/discovery/views.py (3 sites in _search_within_corpus)
- config/graphql/og_metadata_queries.py
- config/graphql/extract_mutations.py
- config/graphql/document_relationship_mutations.py

Corpus.get_documents() now emits a DeprecationWarning. Internal/task
callers (badge_tasks, corpus_tasks, utils/sharing, utils/analyzer,
framework-private LLM helpers in pii/annotations/document_indexing)
migrated to corpus._get_active_documents() to opt out cleanly. The
CAML article paths (caml_article.py, graphql/corpus_types.py) need an
include_caml-aware service method and are tracked under follow-up issues.

Tests added:
- test_document_service.py — structural / partition / back-compat
- test_corpus_objs_service.py — new convenience methods, IDOR safety,
  include_deleted variants, the MCP regression scenario
- test_corpus_get_documents_deprecation.py — warning emitted, internal
  helper silent, parity between the two methods

CLAUDE.md item 7 declares the service layer as the canonical entry point
for any user-context code touching documents.
