+++ config/graphql/document_relationship_mutations.py
@@ -378,9 +378,17 @@ class UpdateDocumentRelationship(graphene.Mutation):
                             message="Corpus not found",
                         )
 
-                    # Validate both documents are in the new corpus
+                    # Validate both documents are in the new corpus.
+                    # Routes through the canonical service so corpus READ is
+                    # enforced against the requesting user.
+                    from opencontractserver.corpuses.corpus_objs_service import (
+                        CorpusObjsService,
+                    )
+
                     docs_in_corpus = (
-                        corpus.get_documents()
+                        CorpusObjsService.get_corpus_documents(
+                            user=info.context.user, corpus=corpus
+                        )
                         .filter(
                             id__in=[
                                 doc_relationship.source_document_id,
