+++ config/graphql/extract_mutations.py
@@ -794,8 +794,19 @@ class CreateExtract(graphene.Mutation):
         extract.save()
 
         if corpus is not None:
-            # Use new DocumentPath-based method to get active documents in corpus
-            extract.documents.add(*corpus.get_documents())
+            # Route through the canonical service so corpus READ is enforced
+            # against the requesting user before the mass-add (the create
+            # mutation already gated on corpus access upstream; this just
+            # keeps the data path through one entry point).
+            from opencontractserver.corpuses.corpus_objs_service import (
+                CorpusObjsService,
+            )
+
+            extract.documents.add(
+                *CorpusObjsService.get_corpus_documents(
+                    user=info.context.user, corpus=corpus
+                )
+            )
         else:
             logger.info("Corpus IS still None... no docs to add.")
